move kernel client tests out of composite kernel tests
