+++ WorkspaceServer.Tests/Kernel/CompositeKernelTests.cs
@@ -199,82 +199,5 @@ x"));
                 .Should()
                 .Be("hello!");
         }
-
-        [Fact]
-        public async Task Kernel_client_surfaces_json_errors()
-        {
-            var kernel = new CompositeKernel
-            {
-                new CSharpKernel(),
-                new FakeKernel("fake")
-                {
-                     Handle = context => Task.CompletedTask
-                }
-            };
-
-            var input = new MemoryStream();
-            var writer = new StreamWriter(input, Encoding.UTF8);
-            writer.WriteLine("{ hello");
-            writer.WriteMessage(new Quit());
-            writer.Flush();
-
-            input.Position = 0;
-
-            var output = new MemoryStream();
-
-            var streamKernel = new KernelStreamClient(kernel,
-                new StreamReader(input),
-                new StreamWriter(output));
-
-            var task = streamKernel.Start();
-            await task;
-
-            output.Position = 0;
-            var reader = new StreamReader(output, Encoding.UTF8);
-
-            var text = reader.ReadToEnd();
-            var events = text.Split(new [] {Environment.NewLine}, StringSplitOptions.RemoveEmptyEntries)
-                .Select(JsonConvert.DeserializeObject<StreamKernelEvent>);
-
-            events.Should().Contain(e => e.EventType == "CommandParseFailure");
-        }
-
-        [Fact]
-        public async Task Kernel_can_pound_r_nuget_using_kernel_client()
-        {
-              var kernel = new CompositeKernel
-            {
-                new CSharpKernel().UseNugetDirective(),
-            };
-
-            var input = new MemoryStream();
-            var writer = new StreamWriter(input, Encoding.UTF8);
-            writer.WriteMessage(new SubmitCode(@"#r ""nuget:Microsoft.Spark, 0.4.0"""));
-            writer.WriteMessage(new Quit());
-
-            input.Position = 0;
-
-            var output = new MemoryStream();
-
-            var streamKernel = new KernelStreamClient(kernel,
-                new StreamReader(input),
-                new StreamWriter(output));
-
-            var task = streamKernel.Start();
-            await task;
-
-            output.Position = 0;
-            var reader = new StreamReader(output, Encoding.UTF8);
-
-            var text = reader.ReadToEnd();
-            var events = text.Split(new[] { Environment.NewLine }, StringSplitOptions.RemoveEmptyEntries)
-                 .Select(JsonConvert.DeserializeObject<StreamKernelEvent>).ToList();
-
-            events.Should()
-                .Contain(e => e.EventType == nameof(NuGetPackageAdded));
-
-            events.Should()
-                 .NotContain(e => e.EventType == nameof(CommandNotRecognized));
-        }
     }
 }
